feat: 게시물 컴포넌트 분리해서 재구현
