+++ frontend/src/styles/pages/Free.css
@@ -39,39 +39,3 @@
     display: none;
   }
 }
-
-.freeboard-wrap {
-  margin-top: 27px;
-  margin-top: 9px;
-}
-
-.post {
-  margin: 0 auto;
-  width: 80%;
-  border: 1px solid #333;
-  border-radius: 5px;
-  padding: 15px 0; /* Adjusted padding */
-  margin-bottom: 9px;
-}
-
-.row {
-  display: flex;
-  align-items: center;
-}
-
-.number,
-.title,
-.content {
-  color: #000;
-  text-align: center;
-  font-family: 'DM Serif Display';
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400px;
-  line-height: 27px; /* 168.75% */
-}
-
-.number,
-.content {
-  margin: 0 16px 0 17px;
-}
